feat(http-server-options): let library users make trust proxy and secure cookies determination
Previously, we made the decision to automatically set trust proxy to true and to use secure cookies
once we determine that the app is running in production. However, not every user may want these
default settings even if their app is running in production. Also, some users may want enable or
disable these settings even if their app is running in an environment other than production, e.g.,
stagin. This commit removes the auto-determination and lets the library user determine whether or
not to trust proxy or to serve cookies only via secure HTTP.